Use u8 values for key/value

+++ src/huffman.rs
@@ -16,10 +16,10 @@ impl Model for Huffman {
     fn sentinel(&self) -> usize {
         *self.length.iter().max().expect("Can not find maximum value.")
     }
-    fn to_btreemap(&self) -> BTreeMap<usize, (usize, usize)> {
-        let mut result: BTreeMap<usize,(usize,usize)> = BTreeMap::new();
+    fn to_btreemap(&self) -> BTreeMap<usize, (u8, u8)> {
+        let mut result: BTreeMap<usize,(u8,u8)> = BTreeMap::new();
         for (sym,&code) in self.codewords.iter().enumerate().filter(|(ix,_)| self.length[*ix as usize] != 0) {
-            result.insert(code << (self.sentinel() - self.length[sym]), (sym, self.length[sym]));
+            result.insert(code << (self.sentinel() - self.length[sym]), (sym as u8, self.length[sym] as u8));
         }
         result
     }
